+++ client/src/lib/i18n.th.ts
@@ -1,0 +1,877 @@
+export const TH_TRANSLATIONS: Record<string, string> = {
+  // Common
+  'app.name': 'หมากรุก',
+  'app.tagline': 'หมากรุกไทยออนไลน์',
+  'nav.play': 'เล่น',
+  'nav.watch': 'ชมสด',
+  'nav.lessons': 'บทเรียน',
+  'nav.puzzles': 'ปริศนา',
+  'nav.games': 'เกม',
+  'nav.about': 'เกี่ยวกับ',
+  'common.back_home': 'กลับหน้าแรก',
+  'common.white': 'ขาว',
+  'common.black': 'ดำ',
+  'common.draw': 'เสมอ',
+  'common.you': 'คุณ',
+  'common.guest': 'ผู้เล่นทั่วไป',
+  'common.anonymous': 'ไม่ระบุตัวตน',
+  'common.cancel': 'ยกเลิก',
+  'common.send': 'ส่ง',
+  'common.sending': 'กำลังส่ง...',
+  'common.close': 'ปิด',
+  'common.retry': 'ลองใหม่',
+  'common.new_game': 'เกมใหม่',
+  'common.loading': 'กำลังโหลด...',
+  'footer.tagline': 'ThaiChess — ฟรีและโอเพนซอร์ส',
+  'footer.inspired': 'ได้แรงบันดาลใจจาก',
+  'footer.community': 'ชุมชน',
+  'footer.github': 'GitHub',
+  'footer.star_github': 'ติดดาวบน GitHub',
+
+  // Appearance
+  'appearance.title': 'กระดานและตัวหมาก',
+  'appearance.board_and_pieces': 'กระดานและตัวหมาก',
+  'appearance.subtitle': 'เลือกธีมกระดานและธีมสีตัวหมากที่อ่านง่ายครั้งเดียว แล้วใช้ได้ทั่วทั้งเว็บ ทั้งออนไลน์ บอท เล่นในเครื่อง และปริศนา',
+  'appearance.open': 'กระดานและตัวหมาก',
+  'appearance.open_short': 'ธีม',
+  'appearance.boards_tab': 'กระดาน',
+  'appearance.colors_tab': 'สีตัวหมาก',
+  'appearance.live_preview': 'พรีวิวสด',
+  'appearance.preview_title': 'ตัวอย่าง',
+  'appearance.preview_subtitle': 'เลื่อนเมาส์เพื่อดูตัวอย่าง คลิกเพื่อใช้งานทันที ระบบจะบันทึกไว้ในอุปกรณ์นี้อัตโนมัติ',
+  'appearance.preview_board': 'ธีมกระดาน',
+  'appearance.preview_piece_theme': 'ธีมสีตัวหมาก',
+  'appearance.core_shape_label': 'รูปทรงหลัก',
+  'appearance.core_shape_desc': 'เป็นรูปทรงเดียวที่รองรับในการเล่นจริง เพื่อให้มองออกเร็วและสม่ำเสมอทุกเกม',
+  'appearance.saved_note': 'บันทึกอัตโนมัติในอุปกรณ์นี้',
+  'appearance.category_classic': 'คลาสสิก',
+  'appearance.category_soft': 'นุ่มนวล',
+  'appearance.category_dark': 'โหมดกลางคืน',
+  'appearance.category_elegant': 'เรียบหรู',
+  'appearance.readability_checked': 'ผ่านการเช็กความชัด',
+  'appearance.grid_contrast': 'คอนทราสต์เส้นตาราง',
+  'appearance.piece_contrast': 'คอนทราสต์ตัวหมากต่ำสุด',
+  'appearance.single_surface_label': 'กระดานสีเดียวแบบหมากรุกไทย',
+  'appearance.single_surface_desc': 'ธีมจะเปลี่ยนเฉพาะโทนกระดาน ผิวสัมผัส และเส้นตารางเท่านั้น',
+  'appearance.comparison_title': 'แบบเดิมเทียบกับหมากรุกไทย',
+  'appearance.comparison_desc': 'ธีมจะไม่สลับสีช่องอีกต่อไป แต่จะเปลี่ยนเฉพาะวัสดุและเส้นตารางของกระดาน',
+  'appearance.comparison_before': 'ก่อนปรับ',
+  'appearance.comparison_after': 'หมากรุกไทยตอนนี้',
+
+  // Home Page
+  'home.hero_title': 'เล่นหมากรุกได้ทันที',
+  'home.hero_desc': 'ไม่ต้องสมัคร เริ่มเกมหมากรุกไทยได้ในไม่กี่วินาที เล่นกับเพื่อน หรือฝึกกับบอท',
+  'home.quick_play': 'เล่นเลย',
+  'home.quick_play_desc': 'จับคู่ทันทีสำหรับเกมสบาย ๆ หรือเกมจัดอันดับ',
+  'home.find_opponent': 'หาคู่แข่ง',
+  'home.no_signup': 'ไม่ต้องสมัคร',
+  'home.free_to_play': 'เล่นฟรี',
+  'home.games_played': 'เล่นแล้ว {count} เกม',
+  'home.play_friend': 'เล่นกับเพื่อน',
+  'home.play_friend_desc': 'แชร์ลิงก์',
+  'home.play_bot': 'เล่นกับบอท',
+  'home.play_bot_desc': '10 บุคลิกบอทธีมเฉพาะ',
+  'home.play_bot_long_desc': 'ท้าดวลกับคู่แข่งหมากรุกไทยที่มีเอกลักษณ์ ชื่อ เรตติ้ง และสไตล์การเล่นต่างกันชัดเจน',
+  'home.puzzles': 'ปริศนา',
+  'home.puzzles_desc': 'ฝึกกลยุทธ์',
+  'home.puzzles_long_desc': 'ฝึกแทคติกด้วยปริศนาที่คัดมาจากเกมจริง',
+  'home.lessons': 'บทเรียน',
+  'home.lessons_desc': 'คอร์สหมากรุกไทยแบบเป็นลำดับ',
+  'home.create_private': 'สร้างเกมส่วนตัว',
+  'home.private_desc': 'เลือกเวลา เลือกสี แล้วแชร์ลิงก์เกมให้เพื่อน',
+  'home.time_control': 'เวลาเล่น',
+  'home.choose_color': 'เลือกสี',
+  'home.color_random': 'สุ่ม',
+  'home.color_white': 'ขาว',
+  'home.color_black': 'ดำ',
+  'home.play_with_friend': 'เล่นกับเพื่อน',
+  'home.creating': 'กำลังสร้าง...',
+  'home.or': 'หรือ',
+  'home.play_local': 'เล่นในเครื่อง (จอเดียวกัน)',
+  'home.play_local_desc': 'สลับกันเดินบนกระดานเดียวในหน้าจอเดียว',
+  'home.watch_live': 'ชมเกมสด',
+  'home.watch_live_desc': 'ดูเกมหมากรุกไทยสาธารณะที่กำลังเล่น และเข้าโหมดผู้ชมได้ทันที',
+  'home.live_now_title': 'กำลังเล่นตอนนี้',
+  'home.live_now_desc': 'เกมสาธารณะที่กำลังเล่นจะอัปเดตแบบเรียลไทม์ เปิดกระดานไหนก็ได้ในโหมดผู้ชมแบบอ่านอย่างเดียว',
+  'home.view_all_live': 'ดูเกมสดทั้งหมด',
+  'home.no_live_games': 'ตอนนี้ยังไม่มีเกมสด',
+  'home.no_live_games_desc': 'กลับมาดูอีกครั้งเมื่อมีเกมสาธารณะกำลังเล่น',
+  'home.join_prompt': 'มีรหัสเกม?',
+  'home.join_link': 'เข้าร่วมเกม',
+  'home.join_title': 'เข้าร่วมเกม',
+  'home.join_desc': 'เปิดเกมที่เพื่อนแชร์มาได้ทันทีด้วยรหัสเกม',
+  'home.join_placeholder': 'ใส่รหัสเกม...',
+  'home.join': 'เข้าร่วม',
+  'home.training_start': 'เริ่มชุดปริศนา',
+  'home.training_continue': 'ฝึกต่อ',
+  'home.training_progress': 'สำเร็จแล้ว {done}/{total} ปริศนา',
+  'home.training_focus': 'ธีมที่เด่นสุดตอนนี้: {theme}',
+  'home.training_resume': 'เล่นล่าสุด: {title}',
+  'home.training_recent': 'แก้ล่าสุด: {title}',
+  'home.streak_start': 'เริ่มสตรีค',
+  'home.streak_continue': 'ลุยต่อ',
+  'home.streak_title': 'สตรีคปริศนา',
+  'home.streak_progress': 'แก้บทเรียนแล้ว {done}/{total} ข้อ',
+  'home.streak_focus': 'ธีมบทเรียนที่เด่นตอนนี้: {theme}',
+  'home.streak_resume': 'บทเรียนล่าสุดที่เล่น: {title}',
+  'home.streak_recent': 'บทเรียนล่าสุดที่แก้ได้: {title}',
+  'home.learn_eyebrow': 'คู่มือเริ่มต้น',
+  'home.learn_title': 'เริ่มจากหน้าที่อ่านแล้วเข้าใจจริง',
+  'home.learn_desc': 'ถ้าคุณเพิ่งเข้ามาใหม่ 3 หน้านี้จะพาคุณจากรู้จักเกม ไปจนถึงเริ่มเล่นได้จริง',
+  'home.learn_card.what_is_title': 'หมากรุกไทยคืออะไร',
+  'home.learn_card.what_is_desc': 'ดูภาพรวมของเกมก่อน ว่าทำไมหมากรุกไทยถึงต่างจากหมากรุกสากล',
+  'home.learn_card.how_to_title': 'วิธีเล่นหมากรุกไทย',
+  'home.learn_card.how_to_desc': 'เรียนการเดินหมาก การหงาย และกฎการนับแบบเข้าใจง่าย',
+  'home.learn_card.play_online_title': 'เล่นหมากรุกไทยออนไลน์',
+  'home.learn_card.play_online_desc': 'ดูว่าเริ่มจากบอท โจทย์ หรือเกมคนจริงแบบไหนเหมาะที่สุด',
+  'lessons.course.eyebrow': 'บทเรียนแบบเป็นลำดับ',
+  'lessons.course.title': 'เรียนหมากรุกไทยเหมือนมีโค้ชคอยพาไปทีละขั้น',
+  'lessons.course.desc': 'ค่อย ๆ เรียนตามโมดูลระดับเริ่มต้น ปานกลาง และขั้นสูงตามลำดับ แต่ละบทจะอธิบายแนวคิดหนึ่งอย่าง แสดงบนกระดาน พาคุณทำทีละขั้น แล้วให้ลองฝึกเองทันที',
+  'lessons.course.stats.progress': 'ความคืบหน้า',
+  'lessons.course.stats.completed': 'บทเรียนที่เรียนจบแล้ว',
+  'lessons.course.stats.shape': 'โครงคอร์ส',
+  'lessons.course.stats.shape_desc': '3 โมดูล ตั้งแต่พื้นฐานไปจนถึงการวางแผน',
+  'lessons.course.stats.format': 'รูปแบบ',
+  'lessons.course.stats.format_value': 'มีโค้ชนำทาง',
+  'lessons.course.stats.format_desc': 'อธิบาย สาธิต ฝึก และตอกย้ำ',
+  'lessons.course.progress_label': 'ความคืบหน้าของคอร์ส',
+  'lessons.course.progress_value': 'สำเร็จแล้ว {percent}%',
+  'lessons.course.next_lesson': 'บทเรียนถัดไป',
+  'lessons.course.minutes': '{minutes} นาที',
+  'lessons.course.start_lesson': 'เริ่มบทเรียน',
+  'lessons.course.complete_title': 'เรียนครบคอร์สแล้ว',
+  'lessons.course.complete_desc': 'คุณเรียนคอร์สหมากรุกไทยชุดปัจจุบันครบแล้ว กลับไปทบทวนบทไหนก็ได้เพื่อให้แนวคิดแม่นขึ้น แล้วค่อยไปต่อที่ปริศนาเพื่อฝึกแพตเทิร์นให้คม',
+  'lessons.course.go_to_puzzles': 'ไปที่ปริศนา',
+  'lessons.course.module_progress': '{done}/{total} บทเสร็จแล้ว',
+  'lessons.course.lesson_label': 'บทเรียน {order}',
+  'lessons.course.status.completed': 'จบแล้ว',
+  'lessons.course.status.resume': 'ทำต่อ',
+  'lessons.course.status.unlocked': 'ปลดล็อกแล้ว',
+  'lessons.course.status.locked': 'ล็อกอยู่',
+  'lessons.player.not_found': 'ไม่พบบทเรียน',
+  'lessons.player.back_to_course': 'กลับไปคอร์ส',
+  'lessons.player.default_correct': 'ถูกต้อง',
+  'lessons.player.default_wrong': 'ตาเดินนี้ไม่ตรงกับเป้าหมายของบทเรียน',
+  'lessons.player.practice_task': 'แบบฝึกหัด',
+  'lessons.player.guided_step': 'ขั้นนำทาง',
+  'lessons.player.mini_practice': 'ฝึกสั้น',
+  'lessons.player.status.completed': 'จบแล้ว',
+  'lessons.player.status.guided': 'นำทาง',
+  'lessons.player.status.practice': 'ฝึก',
+  'lessons.player.header_subtitle': 'บทเรียน · {title}',
+  'lessons.player.course_path': 'เส้นทางคอร์ส',
+  'lessons.player.level_and_order': '{level} · บทเรียน {order}',
+  'lessons.player.counting_aware': 'บทเรียนที่เกี่ยวกับการนับ',
+  'lessons.player.no_counting_dependency': 'ไม่ขึ้นกับกฎการนับ',
+  'lessons.player.lesson_flow': 'ลำดับบทเรียน',
+  'lessons.player.step_of': 'ขั้น {current} จาก {total}',
+  'lessons.player.task_of': 'แบบฝึก {current} จาก {total}',
+  'lessons.player.concept_explanation': 'คำอธิบายแนวคิด',
+  'lessons.player.rule_impact': 'ผลของกฎ',
+  'lessons.player.counting_changes_evaluation': 'กฎการนับมีผลต่อการประเมิน',
+  'lessons.player.counting_not_relevant': 'กฎการนับไม่กระทบบทเรียนนี้',
+  'lessons.player.why_this_move_works': 'ทำไมตานี้จึงใช้ได้ผล',
+  'lessons.player.problem': 'ปัญหา',
+  'lessons.player.what_it_fixes': 'ตานี้แก้อะไร',
+  'lessons.player.new_threat': 'ภัยคุกคามใหม่',
+  'lessons.player.after_the_move': 'หลังจากเดินแล้ว',
+  'lessons.player.compare_moves': 'เปรียบเทียบตาเดิน',
+  'lessons.player.best_move': 'ตาที่ดีที่สุด',
+  'lessons.player.tempting_move': 'ตาที่น่าลองแต่ยังไม่ดีที่สุด',
+  'lessons.player.step_badge': 'ขั้น {number}',
+  'lessons.player.task_badge': 'แบบฝึก {number}',
+  'lessons.player.start_mini_practice': 'เริ่มฝึกสั้น',
+  'lessons.player.next_step': 'ขั้นถัดไป',
+  'lessons.player.review_summary': 'ดูสรุป',
+  'lessons.player.next_task': 'แบบฝึกถัดไป',
+  'lessons.player.next_lesson': 'บทเรียนถัดไป',
+  'lessons.player.summary': 'สรุป',
+  'lessons.player.puzzles_reinforce': 'ปริศนาไว้ตอกย้ำบทเรียนนี้',
+  'lessons.player.puzzles_reinforce_desc': 'หลังเรียนจบแล้ว ลองไปฝึกปริศนาธีมที่เกี่ยวข้องเพื่อให้แนวคิดเปลี่ยนจากคำอธิบายเป็นการจำรูปแบบ',
+  'lessons.player.more_puzzles_coming': 'ลิงก์ปริศนาเพิ่มเติมจะมาแสดงตรงนี้เมื่อชุดโจทย์ของแนวคิดนี้ขยายขึ้น',
+  'lessons.player.previous_lesson': 'บทเรียนก่อนหน้า',
+
+  // Time presets
+  'time.bullet': 'บุลเลต',
+  'time.blitz': 'บลิตซ์',
+  'time.rapid': 'ราปิด',
+  'time.classical': 'คลาสสิก',
+
+  // Bot Game
+  'bot.title': 'เล่นกับบอท',
+  'bot.setup_title': 'เล่นกับคอมพิวเตอร์',
+  'bot.setup_desc': 'เลือกบุคลิกบอท ดูสไตล์ ระดับ และพลังฝีมือโดยประมาณ เลือกฝั่ง แล้วเริ่มเกมซ้อมได้เลย หากเลือกสุ่ม ระบบจะตัดสินฝั่งให้ตอนเริ่มเกมเท่านั้น',
+  'bot.selected_bot': 'บอทที่เลือก',
+  'bot.roster': 'รายชื่อบอท',
+  'bot.featured_opponent': 'คู่ต่อสู้เด่น',
+  'bot.opening_preference': 'สไตล์เปิดเกม',
+  'bot.signature_style': 'เอกลักษณ์การเล่น',
+  'bot.tactical_bias': 'แนวโน้มเชิงแท็กติก',
+  'bot.strategic_weakness': 'จุดอ่อนเชิงกลยุทธ์',
+  'bot.dialogue_preview': 'ตัวอย่างบทสนทนา',
+  'bot.difficulty': 'ระดับความยาก',
+  'bot.level_label': 'เลเวล',
+  'bot.level_short': 'เลเวล {level}',
+  'bot.estimated_elo_label': 'ELO โดยประมาณ',
+  'bot.estimated_elo_range': 'ประมาณ {range} ELO',
+  'bot.estimated_elo_note': 'ค่าพลังฝีมือโดยประมาณนี้อ้างอิงจากพฤติกรรมการเล่น ไม่ใช่เรตทางการ',
+  'bot.easy': 'ง่าย',
+  'bot.easy_desc': 'เดินสุ่มกินตัวพื้นฐาน',
+  'bot.medium': 'ปานกลาง',
+  'bot.medium_desc': 'คิดล่วงหน้า 2 ตา',
+  'bot.hard': 'ยาก',
+  'bot.hard_desc': 'คิดล่วงหน้า 4 ตา',
+  'bot.play_as': 'เลือกฝั่ง',
+  'bot.random': 'สุ่ม',
+  'bot.random_assigned': 'กำหนดตอนเริ่มเกม',
+  'bot.side_locked': 'ล็อกฝั่งก่อนเริ่มเกม',
+  'bot.start': 'เริ่มเกม',
+  'bot.thinking': 'กำลังคิด...',
+  'bot.chat_thinking': 'กำลังคิด',
+  'bot.your_turn': 'ตาของคุณ',
+  'bot.bot_thinking': 'บอทกำลังคิด...',
+  'bot.you_won': 'คุณชนะ!',
+  'bot.bot_wins': 'บอทชนะ',
+  'bot.resign': 'ยอมแพ้',
+  'bot.resign_confirm': 'คุณแน่ใจหรือไม่ว่าต้องการยอมแพ้?',
+  'bot.vs_bot': 'กับบอท',
+
+  // Puzzles
+  'puzzle.title': 'สตรีคปริศนา',
+  'puzzle.desc': 'เล่นปริศนาต่อเนื่อง เก็บคะแนน และให้ระบบปรับความยากแบบซ่อนอยู่ตามผลงานของคุณ',
+  'puzzle.completed': '{done}/{total} สำเร็จ',
+  'puzzle.completed_summary': 'ความคืบหน้าทั้งหมด',
+  'puzzle.progress_hint': 'ปริศนาที่แก้แล้วจะถูกจำไว้ในอุปกรณ์นี้',
+  'puzzle.all': 'ทั้งหมด',
+  'puzzle.all_lessons': 'บทเรียนทั้งหมด',
+  'puzzle.beginner': 'เริ่มต้น',
+  'puzzle.intermediate': 'ปานกลาง',
+  'puzzle.advanced': 'ขั้นสูง',
+  'puzzle.filter_label': 'ชุดฝึกตอนนี้',
+  'puzzle.filter_all_desc': 'รวมทุกบทเรียนที่เปิดใช้ไว้ในชุดเดียว',
+  'puzzle.filter_beginner_desc': 'แท็กติกสั้นและรูปจนพื้นฐาน',
+  'puzzle.filter_intermediate_desc': 'แนวคิดสองจังหวะที่ลงโทษเกมรับหลวม',
+  'puzzle.filter_advanced_desc': 'โจทย์คำนวณหนักขึ้นเมื่อมีชุดเพิ่ม',
+  'puzzle.theme_drill_title': 'ฝึกตามธีม',
+  'puzzle.theme_drill_desc': 'เลือกฝึกแนวคิดเดียวซ้ำ ๆ เมื่ออยากซ้อมเฉพาะทาง',
+  'puzzle.theme_drill_count': '{count} ธีมในชุดนี้',
+  'puzzle.theme_all': 'ทุกแนวคิด',
+  'puzzle.track_progress': 'ความคืบหน้าชุด {track}',
+  'puzzle.track_completed': '{done}/{total} ในชุดนี้',
+  'puzzle.remaining_label': 'ยังไม่เคยแก้',
+  'puzzle.remaining_desc': 'โจทย์ใหม่จะถูกจัดไว้ก่อนเสมอ',
+  'puzzle.focus_label': 'ธีมเด่น',
+  'puzzle.focus_desc': 'ช่วยเลือกว่าจะฝึกแนวคิดไหนต่อ',
+  'puzzle.focus_empty': 'ยังไม่มีธีม',
+  'puzzle.next_up': 'บทเรียนถัดไป',
+  'puzzle.next_up_fresh': 'เริ่มจากโจทย์ที่ยังไม่เคยแก้ในชุดนี้ก่อน',
+  'puzzle.next_up_review': 'คุณเคลียร์ชุดนี้แล้ว ลองทบทวนโจทย์เดิมเพื่อให้แม่นขึ้น',
+  'puzzle.start_here': 'เริ่มตรงนี้',
+  'puzzle.practice_title': 'คิวฝึก',
+  'puzzle.practice_desc': 'โจทย์ใหม่จะมาก่อน ส่วนโจทย์ที่แก้แล้วเก็บไว้ทบทวน',
+  'puzzle.new_badge': 'ใหม่',
+  'puzzle.solved_badge': 'แก้แล้ว',
+  'puzzle.empty_title': 'ยังไม่มีปริศนาในชุดนี้',
+  'puzzle.empty_desc': 'ลองเปลี่ยนระดับความยาก หรือกลับมาอีกครั้งเมื่อมีบทเรียนเพิ่ม',
+  'puzzle.empty_theme_title': 'ยังไม่มีโจทย์ธีม {theme} ในชุดนี้',
+  'puzzle.empty_theme_desc': 'ลองเปลี่ยนแนวคิด หรือออกจากชุด {track} เพื่อหาโจทย์ {theme} เพิ่ม',
+  'puzzle.to_move': '{color} เดิน',
+  'puzzle.find_best': 'หาตาเดินที่ดีที่สุดสำหรับฝั่ง{color}!',
+  'puzzle.step': 'ขั้นที่ {current} จาก {total}',
+  'puzzle.correct': 'ถูกต้อง!',
+  'puzzle.solved_hint': 'แก้ได้โดยใช้คำใบ้',
+  'puzzle.solved_clean': 'แก้ได้โดยไม่ใช้คำใบ้!',
+  'puzzle.wrong': 'ไม่ถูกต้อง!',
+  'puzzle.wrong_desc': 'นั่นไม่ใช่ตาเดินที่ดีที่สุด ลองอีกครั้ง!',
+  'puzzle.hint': 'คำใบ้',
+  'puzzle.next': 'ปริศนาถัดไป',
+  'puzzle.previous': 'ก่อนหน้า',
+  'puzzle.all_puzzles': 'ปริศนาทั้งหมด',
+  'puzzle.back_to_lessons': 'กลับไปบทเรียน',
+  'puzzle.not_found': 'ไม่พบปริศนา',
+  'puzzle.back': 'กลับไปปริศนา',
+  'puzzle.lesson': 'บทเรียน',
+  'puzzle.lessons_nav': 'บทเรียน',
+  'puzzle.lessons_eyebrow': 'โหมดบทเรียน',
+  'puzzle.lessons_title': 'บทเรียน',
+  'puzzle.lessons_desc': 'เรียนหมากรุกไทยเป็นคอร์สตามลำดับ แล้วค่อยต่อยอดด้วยปริศนาที่ตรงกับแนวคิดของบทเรียน',
+  'puzzle.lessons_tracks_title': 'เส้นทางการเรียน',
+  'puzzle.lessons_tracks_desc': 'ระดับเริ่มต้น ระดับกลาง และระดับสูง ยังคงอยู่ที่นี่ในฐานะเส้นทางฝึกแบบตั้งใจ',
+  'puzzle.play_streak': 'เล่นสตรีค',
+  'puzzle.streak_nav': 'สตรีคปริศนา',
+  'puzzle.streak_eyebrow': 'โหมดหลัก',
+  'puzzle.streak_title': 'สตรีคปริศนา',
+  'puzzle.streak_desc': 'แก้ปริศนาต่อเนื่อง เก็บสตรีค และให้ระบบค่อย ๆ ปรับความท้าทายขึ้นหรือลงจากการเล่นของคุณ',
+  'puzzle.streak_prompt_title': 'รักษาสตรีคไว้',
+  'puzzle.streak_prompt_desc': 'ตอบถูกแล้วได้คะแนนทันที และโจทย์ถัดไปจะขึ้นต่อแบบลื่นไหล',
+  'puzzle.streak_score_label': 'คะแนน',
+  'puzzle.streak_label': 'สตรีค',
+  'puzzle.streak_session_label': 'แก้ได้',
+  'puzzle.streak_checkpoint_label': 'เช็กพอยต์',
+  'puzzle.streak_checkpoint_progress': '{current}/{total} ถึงสัญญาณถัดไป',
+  'puzzle.streak_flow_label': 'จังหวะ',
+  'puzzle.streak_flow_desc': 'ปรับแบบอัตโนมัติ',
+  'puzzle.streak_puzzle_label': 'โจทย์ {number}',
+  'puzzle.streak_points': '+{points} คะแนน',
+  'puzzle.streak_focus_title': 'โจทย์ตอนนี้',
+  'puzzle.streak_focus_keep': 'ตั้งหลักใหม่ แล้วเริ่มอีกครั้ง',
+  'puzzle.streak_focus_desc': 'โหมดนี้ซ่อนความยากไว้ เพื่อให้เล่นต่อได้เร็วและลื่นที่สุด',
+  'puzzle.streak_session_title': 'เซสชัน',
+  'puzzle.streak_best_label': 'สตรีคสูงสุด',
+  'puzzle.streak_lessons_progress_label': 'บทเรียน',
+  'puzzle.visit_lessons': 'เปิดบทเรียน',
+  'puzzle.streak_end_eyebrow': 'จบเซสชัน',
+  'puzzle.streak_end_title': 'สตรีคจบแล้ว',
+  'puzzle.streak_end_desc': 'พลาดหนึ่งครั้งก็จบสตรีค แต่ความคืบหน้าของบทเรียนยังถูกบันทึกไว้',
+  'puzzle.streak_ended_at': 'สตรีคจบที่ {streak}',
+  'puzzle.streak_end_summary': 'สตรีคสุดท้าย {streak} · คะแนนเซสชัน {score}',
+  'puzzle.retry_puzzle': 'ลองโจทย์เดิมอีกครั้ง',
+  'puzzle.streak_try_again': 'เริ่มสตรีคใหม่',
+  'puzzle.streak_milestone_improving': 'คุณกำลังดีขึ้น เล่นต่อให้ลื่นเลย',
+  'puzzle.streak_milestone_harder': 'ตอนนี้กำลังเจอโจทย์ที่ยากขึ้นแล้ว ระวังให้ดี',
+  'puzzle.source': 'ที่มา',
+  'puzzle.source_real_game_ply': 'เกมจริง · ตาที่ {ply}',
+  'puzzle.source_seed_game_ply': 'เกมตัวอย่าง · ตาที่ {ply}',
+  'puzzle.source_starter_pack': 'ชุดเริ่มต้น',
+  'puzzle.source_curated_tactic': 'แทคติกคัดสรร',
+  'puzzle.source_tactical_motif': 'มอติฟแทคติก',
+  'puzzle.source_review_batch': 'ชุดรอตรวจ',
+  'puzzle.rating_label': 'ระดับที่แนะนำ',
+  'puzzle.rating_short': 'เรต {score}',
+  'puzzle.success_rate': 'อัตราสำเร็จ',
+  'puzzle.attempts_label': 'จำนวนครั้ง',
+  'puzzle.activity_title': 'ความคืบหน้ารอบนี้',
+  'puzzle.activity_status_label': 'สถานะ',
+  'puzzle.activity_status_new': 'ใหม่',
+  'puzzle.activity_status_in_progress': 'กำลังฝึก',
+  'puzzle.activity_status_solved': 'แก้แล้ว',
+  'puzzle.activity_last_played': 'เล่นล่าสุด {date}',
+  'puzzle.activity_completed_on': 'แก้สำเร็จ {date}',
+  'puzzle.related_theme_title': 'ธีมเดียวกันต่อ',
+  'puzzle.related_theme_desc': 'ต่อยอดธีม {theme} ด้วยโจทย์ถัดไปเหล่านี้',
+  'puzzle.related_theme_empty': 'ยังไม่มีโจทย์ {theme} เพิ่ม',
+
+  // Puzzle themes
+  'theme.BasicCheckmate': 'รูปจนพื้นฐาน',
+  'theme.BackRank': 'รุกจนแถวหลัง',
+  'theme.Checkmate': 'รุกจน',
+  'theme.Clearance': 'เปิดทาง',
+  'theme.Decoy': 'ล่อ',
+  'theme.Defense': 'ตั้งรับ',
+  'theme.Deflection': 'เบี่ยงตัวป้องกัน',
+  'theme.DoubleAttack': 'โจมตีสองเป้า',
+  'theme.DoubleCheck': 'รุกสองทาง',
+  'theme.ExchangeSacrifice': 'สังเวยเรือแลกตัว',
+  'theme.HangingPiece': 'ตัวลอย',
+  'theme.Interference': 'ขวางแนว',
+  'theme.MateIn1': 'จนใน 1',
+  'theme.MateIn2': 'จนใน 2',
+  'theme.MateIn3': 'จนใน 3+',
+  'theme.MatingNet': 'ข่ายจน',
+  'theme.Overloading': 'รับหลายหน้าที่',
+  'theme.PerpetualCheck': 'รุกไล่ไม่จบ',
+  'theme.Fork': 'โจมตีซ้อน',
+  'theme.Pin': 'ตรึง',
+  'theme.Promotion': 'หงายเบี้ย',
+  'theme.Discovery': 'เปิดแนว',
+  'theme.RemovalOfDefender': 'กำจัดตัวป้องกัน',
+  'theme.Sacrifice': 'สังเวย',
+  'theme.Simplification': 'ผ่อนเข้าสู่ชนะง่าย',
+  'theme.SmotheredMate': 'ม้าจนปิดล้อม',
+  'theme.Skewer': 'เสียบ',
+  'theme.Stalemate': 'อับจนเสมอ',
+  'theme.SupportMate': 'จนมีตัวค้ำ',
+  'theme.Tactic': 'กลยุทธ์',
+  'theme.TrappedPiece': 'ตัวติดกับ',
+  'theme.VulnerableKing': 'กษัตริย์เปราะบาง',
+  'theme.Windmill': 'กังหันกิน',
+  'theme.XRay': 'โจมตีทะลุแนว',
+  'theme.Zwischenzug': 'แทรกหมาก',
+  'theme.Endgame': 'เกมจบ',
+  'theme.Desperado': 'ตายก่อนขอกิน',
+
+  // Quick Play / Matchmaking
+  'quick.title': 'เล่นด่วน',
+  'quick.desc': 'หาคู่แข่งทันที ไม่ต้องแชร์ลิงก์!',
+  'quick.searching': 'กำลังหาคู่แข่ง...',
+  'quick.search_time': 'กำลังค้นหามา {time}',
+  'quick.queue': '{count} ผู้เล่นในคิว',
+  'quick.find': 'หาคู่แข่ง',
+  'quick.rated_available': 'จัดอันดับได้',
+  'quick.rated_unavailable': 'ปิดเกมจัดอันดับ',
+  'quick.casual_only': 'เกมทั่วไป',
+  'quick.load_failed': 'ไม่สามารถเริ่มเล่นด่วนได้ในขณะนี้',
+  'quick.rated_signed_in': 'จะเป็นเกมจัดอันดับเมื่อเจอคู่แข่งที่เข้าสู่ระบบเช่นกัน',
+  'quick.rated_restricted': 'บัญชีนี้ยังเล่นควิกเพลย์แบบทั่วไปได้ แต่ปิดการจับคู่แบบจัดอันดับไว้',
+  'quick.rated_sign_in': 'เข้าสู่ระบบเพื่อปลดล็อกเกมจัดอันดับ',
+
+  // Game Page
+  'game.waiting_title': 'รอคู่แข่ง',
+  'game.waiting_desc': 'แชร์ลิงก์นี้ให้เพื่อนเพื่อเริ่มเล่น',
+  'game.copy': 'คัดลอก',
+  'game.copied': 'คัดลอกแล้ว!',
+  'game.playing_as': 'คุณเล่นฝั่ง{color}',
+  'game.rated': 'จัดอันดับ',
+  'game.casual': 'ทั่วไป',
+  'game.rating_change': 'เรตติ้ง',
+  'game.connecting': 'กำลังเชื่อมต่อ...',
+  'game.load_failed': 'ไม่สามารถโหลดเกมนี้ได้ในขณะนี้',
+  'game.error': 'ข้อผิดพลาด',
+  'game.opponent': 'คู่แข่ง',
+  'game.online': 'ออนไลน์',
+  'game.offline': 'ออฟไลน์',
+  'game.active_now': 'กำลังเล่น',
+  'game.idle': 'ไม่ได้ใช้งาน',
+  'game.away': 'อยู่นอกหน้าเกม',
+  'game.ping_short': 'ปิง',
+  'game.ping_unknown': 'ไม่มีข้อมูลปิง',
+  'game.ping_value': 'ค่าหน่วง {ms} ms',
+  'game.to_move': 'ถึงตา',
+  'game.your_turn': 'ตาของคุณ',
+  'game.opponent_turn': 'ตาฝ่ายตรงข้าม',
+  'game.you_won': 'คุณชนะ!',
+  'game.you_lost': 'คุณแพ้',
+  'game.draw_offer_received': 'คู่แข่งเสนอเสมอ',
+  'game.accept': 'ยอมรับ',
+  'game.decline': 'ปฏิเสธ',
+  'game.offer_draw': '½ เสมอ',
+  'game.resign': '⚐ ยอมแพ้',
+  'game.resign_confirm': 'คุณแน่ใจหรือไม่ว่าต้องการยอมแพ้?',
+  'game.share': 'แชร์',
+  'game.open_spectator': 'เปิดมุมมองผู้ชม',
+  'game.piece_guide': '📖 คู่มือตัวหมาก',
+  'game.opponent_dc': 'คู่แข่งหลุดออก กำลังรอเชื่อมต่อใหม่...',
+  'game.game_label': 'เกม:',
+  'game.counting_title': 'กฎการนับ',
+  'game.counting_board_honor': 'การนับแบบหมดเบี้ย',
+  'game.counting_pieces_honor': 'การนับแบบขุนเปล่า',
+  'game.counting_status': '{type}: ฝั่ง{color} นับอยู่ที่ {current}/{limit}',
+  'game.counting_final': '{type}: ตาโจมตีสุดท้ายก่อนเสมอ',
+  'game.counting_available': '{type}: ฝั่ง{color} สามารถเริ่มนับได้',
+  'game.counting_start': 'เริ่มนับ',
+  'game.counting_stop': 'หยุดนับ',
+  'game.premove_set': 'ตั้งตาเดินล่วงหน้าแล้ว',
+  'game.check_status': 'กำลังถูกรุก',
+  'game.nav_hint': 'ใช้ปุ่มลูกศรเพื่อดูตาเดิน',
+  'game.select_piece_style': 'เลือกรูปแบบตัวหมาก',
+  'game.piece_style': 'ตัวหมาก',
+  'game.piece_style_current': 'ปัจจุบัน',
+  'game.piece_style_western': 'สากล',
+  'game.piece_style_makruk': 'หมากรุก',
+  'game.match_status': 'สถานะเกม',
+  'game.playing_as_label': 'คุณเล่นฝั่ง',
+  'game.review_mode': 'มุมมองกระดาน',
+  'game.reviewing_history': 'กำลังดูประวัติเดิน',
+  'game.live_position': 'ตำแหน่งปัจจุบัน',
+  'game.return_to_live': 'กลับไปตาเดินปัจจุบัน',
+  'game.live_match_title': 'เกมหมากรุกสด',
+  'game.live_match_desc': 'ให้กระดานเป็นจุดสำคัญที่สุด ส่วนเวลา การทบทวน และคำสั่งควรอยู่ใกล้มือโดยไม่แย่งความสนใจ',
+  'game.spectator_mode': 'โหมดผู้ชม',
+  'game.watching_live_game': 'กำลังชมเกมสด',
+  'game.spectator_desc': 'มุมมองแบบอ่านอย่างเดียว พร้อมกระดานสด นาฬิกา และรายการตาเดิน',
+  'game.live_now': 'สด',
+  'game.final_position': 'ตำแหน่งสุดท้าย',
+  'game.read_only': 'อ่านอย่างเดียว',
+  'game.flip_board': 'สลับมุมกระดาน',
+  'game.white_perspective': 'มุมมองฝ่ายขาว',
+  'game.black_perspective': 'มุมมองฝ่ายดำ',
+  'game.live_position_locked': 'ระหว่างเกม กระดานจะแสดงตำแหน่งสดล่าสุดเสมอ',
+  'game.spectator_turn': 'ถึงตาฝั่ง{color}',
+  'game.spectator_controls_hidden': 'ซ่อนคำสั่งสำหรับผู้เล่นในโหมดผู้ชม',
+  'game.post_game_title': 'ทบทวนเกมที่จบแล้ว',
+  'game.post_game_desc': 'ไล่ดูตาเดิน ศึกษาจุดจบของเกม หรือไปวิเคราะห์ต่อได้ทันที',
+  'game.side_title': 'แผงข้อมูลเกม',
+  'game.you_label': 'คุณ',
+  'game.reviewing_position': 'เกมจบแล้ว',
+
+  // Game Over Modal
+  'gameover.you_win': 'คุณชนะ!',
+  'gameover.you_lost': 'คุณแพ้',
+  'gameover.draw': 'เสมอ',
+  'gameover.by_checkmate': 'โดยรุกจน',
+  'gameover.by_resign': 'โดยยอมแพ้',
+  'gameover.by_timeout': 'หมดเวลา',
+  'gameover.by_stalemate': 'โดยอับ',
+  'gameover.by_agreement': 'ตกลงเสมอ',
+  'gameover.by_material': 'ตัวหมากไม่เพียงพอ',
+  'gameover.by_counting': 'โดยกฎการนับ',
+  'gameover.by_unknown': 'เกมจบแล้ว',
+  'gameover.rematch': 'แข่งอีกครั้ง',
+  'gameover.rematch_accept': 'ยอมรับแข่งอีกครั้ง',
+  'gameover.rematch_sent': 'ส่งคำขอแล้ว',
+  'gameover.rematch_waiting': 'ส่งคำขอแข่งอีกครั้งแล้ว กำลังรอคู่แข่งตอบรับ',
+  'gameover.rematch_incoming': 'คู่แข่งต้องการแข่งอีกครั้ง',
+  'gameover.is_victorious': 'ชนะ',
+
+  // Local Game
+  'local.title': 'เล่นในเครื่อง',
+  'local.view_as': 'มุมมอง:',
+  'local.turn': 'ตาฝั่ง{color}',
+  'local.wins': 'ฝั่ง{color}ชนะ!',
+  'local.play_online': 'เล่นออนไลน์',
+
+  // Move History
+  'moves.title': 'ตาเดิน',
+  'moves.empty': 'ยังไม่มีตาเดิน',
+  'moves.first_position': 'ตำแหน่งแรก',
+  'moves.previous_move': 'ตาก่อนหน้า',
+  'moves.next_move': 'ตาถัดไป',
+  'moves.last_move': 'ตำแหน่งล่าสุด',
+
+  // Piece Guide
+  'guide.title': 'คู่มือตัวหมาก',
+  'guide.king': 'ขุน',
+  'guide.queen': 'เม็ด',
+  'guide.bishop': 'โคน',
+  'guide.rook': 'เรือ',
+  'guide.knight': 'ม้า',
+  'guide.pawn': 'เบี้ย',
+  'guide.promoted': 'เบี้ยหงาย',
+  'guide.king_move': 'เดินได้ 1 ช่องทุกทิศทาง',
+  'guide.queen_move': 'เดินได้ 1 ช่องแนวทแยง',
+  'guide.bishop_move': 'เดินได้ 1 ช่องแนวทแยงหรือ 1 ช่องหน้า',
+  'guide.rook_move': 'เดินได้ไม่จำกัดในแนวตั้งและแนวนอน',
+  'guide.knight_move': 'รูปตัว L: 2+1 ช่อง',
+  'guide.pawn_move': 'เดินหน้า 1 ช่อง; กินแนวทแยง',
+  'guide.promoted_move': 'เหมือนเม็ด เดินทแยง 1 ช่อง',
+
+  // Connection
+  'conn.connected': 'เชื่อมต่อแล้ว',
+  'conn.reconnecting': 'กำลังเชื่อมต่อใหม่...',
+  'conn.disconnected': 'หลุดการเชื่อมต่อ — กำลังเชื่อมต่อใหม่...',
+
+  // About Page
+  'about.eyebrow': 'เรียนรู้หมากรุก',
+  'about.title': 'เกี่ยวกับหมากรุกไทย',
+  'about.hero_desc': 'หมากรุกไทยคือหมากรุกดั้งเดิมของไทย จังหวะเกมค่อยเป็นค่อยไปกว่าเดิม ลดการแลกง่าย ๆ และเต็มไปด้วยรูปแบบตัวหมากที่ไม่เหมือนหมากรุกสากล เว็บไซต์นี้ถูกสร้างขึ้นเพื่อให้คนเรียนรู้และเล่นออนไลน์ได้ง่ายขึ้น',
+  'about.cta_play': 'เริ่มเล่น',
+  'about.cta_puzzles': 'ลองปริศนา',
+  'about.quick_facts': 'ข้อมูลเร็ว',
+  'about.fact_heritage_label': 'ต้นกำเนิด',
+  'about.fact_heritage_value': 'หมากรุกไทยดั้งเดิม',
+  'about.fact_style_label': 'ลักษณะเกม',
+  'about.fact_style_value': 'รอบคอบ วางแผน ละเอียด',
+  'about.intro_title': 'ทำไมเราสร้างเว็บนี้',
+  'about.intro1': 'หมากรุกไทยเป็นแพลตฟอร์ม<strong>ฟรีและโอเพนซอร์ส</strong>สำหรับเล่นหมากรุกไทยกับทุกคนในโลก ไม่ต้องสมัคร ไม่มีโฆษณา ไม่มีค่าใช้จ่าย — แค่หมากรุกไทยบริสุทธิ์',
+  'about.intro2': 'พันธกิจของเราคือการทำให้<strong>หมากรุกไทยเป็นที่รู้จักทั่วโลก</strong> หมากรุกไทยเป็นหนึ่งในเกมกระดานที่เก่าแก่ที่สุดในโลก ใกล้เคียงกับจตุรงค์ของอินเดียมากกว่าหมากรุกสากล สมควรมีแพลตฟอร์มออนไลน์ระดับโลก',
+  'about.intro3': 'ได้แรงบันดาลใจจาก <a href="https://lichess.org" target="_blank" rel="noopener" class="text-primary hover:text-primary-light underline">Lichess</a> และ <a href="https://chess.com" target="_blank" rel="noopener" class="text-primary hover:text-primary-light underline">Chess.com</a> เราเชื่อว่าแพลตฟอร์มเกมดีๆ ควรเข้าถึงได้ทุกคน',
+  'about.stats_title': 'สถิติแพลตฟอร์ม',
+  'about.games_played': 'เกมที่เล่น',
+  'about.total_moves': 'ตาเดินทั้งหมด',
+  'about.white_wins': 'ฝั่งขาวชนะ',
+  'about.black_wins': 'ฝั่งดำชนะ',
+  'about.what_title': 'หมากรุกไทยคืออะไร?',
+  'about.what1': '<strong>หมากรุกไทย</strong> เป็นหมากรุกดั้งเดิมของไทย พัฒนามาจากจตุรงค์ของอินเดียโบราณ และเล่นกันในไทยมาหลายศตวรรษ',
+  'about.what2': 'ต่างจากหมากรุกสากล หมากรุกไทยยังคงรักษาลักษณะดั้งเดิมไว้ — เม็ด (ควีน) อ่อนแอ โคน (บิชอป) เดินต่างออกไป และเบี้ยเริ่มที่แถวที่สาม ทำให้เกมมีกลยุทธ์ลึกซึ้งและสวยงามเป็นเอกลักษณ์',
+  'about.what3': 'หมากรุกไทยเป็นที่นิยมในไทยและกัมพูชา (เรียกว่า อุกจตุรงค์) ได้รับการยอมรับจากสหพันธ์หมากรุกโลกและมีการแข่งขันระดับนานาชาติ',
+  'about.what4': 'เราเชื่อว่าหมากรุกไทยสมควรได้รับการยอมรับระดับโลกเทียบเท่าหมากรุกสากล ช่วยเราเผยแพร่!',
+  'about.diff_eyebrow': 'อะไรที่ต่างออกไป',
+  'about.diff_title': 'หมากรุกไทยให้ความรู้สึกต่างจากหมากรุกสากลอย่างไร',
+  'about.diff1': 'เม็ดอ่อนกว่าควีนสากลมาก เกมรุกจึงค่อยเป็นค่อยไปและการวางตำแหน่งสำคัญขึ้น',
+  'about.diff2': 'รูปแบบการเดินของโคนเปลี่ยนแนวทแยงและทำให้เกิดแทคติกที่ผู้เล่นหมากรุกสากลต้องเรียนรู้ใหม่',
+  'about.diff3': 'เบี้ยเริ่มบนแถวที่สามและหงายไม่เหมือนเดิม ทำให้โอเพนนิงและเอนด์เกมมีเอกลักษณ์ของตัวเอง',
+  'about.diff4': 'กฎการนับทำให้เอนด์เกมบางแบบมีจังหวะต่างจากหมากรุกสากลและต้องเล่นอย่างแม่นยำ',
+  'about.pieces_eyebrow': 'พื้นฐานตัวหมาก',
+  'about.pieces_title': 'จำการเดินของตัวหมากให้เร็ว',
+  'about.learn_by_playing': 'เรียนรู้ต่อด้วยเกมเร็ว',
+  'about.opensource_title': 'โอเพนซอร์ส',
+  'about.opensource_desc': 'หมากรุกไทยเป็นโอเพนซอร์ส 100% ภายใต้สัญญาอนุญาต MIT และโค้ดทั้งหมดอยู่บน GitHub',
+  'about.footer': 'หมากรุกไทย — ฟรีและโอเพนซอร์ส — สร้างด้วยรักเพื่อหมากรุกไทย',
+
+  // Games Page
+  'games.title': 'เกมล่าสุด',
+  'games.count': 'เล่นแล้ว {count} เกม',
+  'games.empty': 'ยังไม่มีเกม',
+  'games.empty_desc': 'เป็นคนแรกที่เล่น!',
+  'games.start': 'เริ่มเล่น',
+  'games.featured_title': 'เกมที่น่าดูก่อน',
+  'games.featured_desc': 'เกมที่ยาวขึ้นหรือเล่นกันจริงจังมากขึ้นจะถูกดันขึ้นมาก่อน เพื่อให้คุณเปิดวิเคราะห์เกมที่มีสาระได้เร็วขึ้น',
+  'games.no_featured_title': 'หน้านี้มีแต่เกมจบเร็วเป็นส่วนใหญ่',
+  'games.no_featured_desc': 'ผลล่าสุดชุดนี้ส่วนใหญ่จบจากการตกลงเสมอ ยอมแพ้ทันที หรือหมดเวลาเร็วเกินไป',
+  'games.low_signal_title': 'เกมจบเร็วและเกมสัญญาณต่ำ',
+  'games.low_signal_desc': 'เกมเหล่านี้ยังนับอยู่ แต่โดยมากให้บทเรียนได้น้อยกว่า เพราะจบเร็วเกินไป',
+  'games.low_signal_badge': 'จบเร็ว',
+  'games.col_game': 'เกม',
+  'games.col_time': 'เวลา',
+  'games.col_result': 'ผลลัพธ์',
+  'games.col_moves': 'ตาเดิน',
+  'games.col_when': 'เมื่อไหร่',
+  'games.prev': 'ก่อนหน้า',
+  'games.next': 'ถัดไป',
+  'games.page': 'หน้า {current} จาก {total}',
+  'games.view_leaderboard': 'ตารางคะแนน',
+  'games.filter_all': 'ทั้งหมด',
+  'games.filter_rated': 'จัดอันดับ',
+  'games.filter_casual': 'ทั่วไป',
+  'games.filter_bot': 'บอท',
+  'games.bot_badge': 'บอท',
+  'games.bot_games': 'เกมกับบอท',
+  'games.bot_win_rate': 'อัตราชนะกับบอท',
+  'games.bot_highest_level': 'ระดับสูงสุดที่ชนะได้',
+  'games.reason_checkmate': 'รุกจน',
+  'games.reason_resignation': 'ยอมแพ้',
+  'games.reason_timeout': 'หมดเวลา',
+  'games.reason_stalemate': 'อับ',
+  'games.reason_agreement': 'ตกลง',
+  'games.reason_material': 'ตัวหมากไม่พอ',
+  'games.reason_counting': 'กฎการนับ',
+  'games.reason_draw': 'เสมอ',
+  'games.reason_unknown': 'จบเกม',
+  'live.title': 'ชมเกมสด',
+  'live.desc': 'ค้นหาเกมหมากรุกไทยสาธารณะที่กำลังเล่น และตามดูทุกตาเดินในโหมดผู้ชม',
+  'live.live_now': 'กำลังเล่น',
+  'live.live_count': 'เกมสด {count} เกม',
+  'live.recently_finished': 'เพิ่งจบไป',
+  'live.recently_finished_desc': 'เกมสาธารณะเหล่านี้เพิ่งจบ และยังเปิดทบทวนได้',
+  'live.empty_title': 'ตอนนี้ยังไม่มีเกมสด',
+  'live.empty_desc': 'เมื่อมีเกมเล่นด่วนสาธารณะเริ่มขึ้น เกมเหล่านั้นจะปรากฏที่นี่โดยอัตโนมัติ',
+  'live.card_live': 'สด',
+  'live.card_finished': 'จบแล้ว',
+  'live.watchers': 'ผู้ชม {count} คน',
+  'live.watch': 'ชมเกม',
+  'live.view_spectator': 'เปิดโหมดผู้ชม',
+  'live.mode_quick_play': 'เล่นด่วน',
+  'live.mode_rated': 'เล่นด่วนจัดอันดับ',
+  'live.mode_casual': 'เล่นด่วนทั่วไป',
+  'live.moves': '{count} ตาเดิน',
+  'live.updated': 'ตาเดินล่าสุดเมื่อ {n} นาทีที่แล้ว',
+  'live.just_now': 'อัปเดตเมื่อสักครู่',
+
+  // Feedback
+  'feedback.button': 'แจ้งปัญหา',
+  'feedback.title': 'ส่งความคิดเห็น',
+  'feedback.bug': 'บั๊ก',
+  'feedback.feature': 'ฟีเจอร์',
+  'feedback.other': 'อื่นๆ',
+  'feedback.placeholder_bug': 'เกิดอะไรขึ้น? คุณคาดหวังอะไร?',
+  'feedback.placeholder_feature': 'คุณอยากได้ฟีเจอร์อะไร?',
+  'feedback.placeholder_other': 'บอกเราได้เลย...',
+  'feedback.thanks': 'ขอบคุณ!',
+  'feedback.thanks_desc': 'ความคิดเห็นของคุณช่วยพัฒนา ThaiChess ออนไลน์',
+  'feedback.error': 'ส่งไม่สำเร็จ กรุณาลองใหม่',
+
+  // Feedback Messages Page
+  'feedback_page.title': 'ข้อความแจ้งปัญหา',
+  'feedback_page.count': '{count} ข้อความ',
+  'feedback_page.filter_all': 'ทั้งหมด',
+  'feedback_page.empty': 'ยังไม่มีข้อความ',
+  'feedback_page.empty_desc': 'ข้อความจากผู้ใช้จะแสดงที่นี่',
+  'feedback_page.detail_page': 'หน้า',
+  'feedback_page.detail_date': 'วันที่',
+  'feedback_page.loading': 'กำลังโหลดหน้าเครื่องมือดูแล...',
+  'feedback_page.load_failed': 'โหลดข้อความแจ้งปัญหาไม่สำเร็จ',
+  'feedback_page.moderate_failed': 'จัดการข้อความแจ้งปัญหาไม่สำเร็จ',
+  'feedback_page.removed_by_admin': 'ลบโดยแอดมิน',
+  'feedback_page.delete': 'ลบ',
+  'fair_play.report_action': 'รายงานการเล่นต้องสงสัย',
+  'fair_play.report_sent': 'ส่งรายงานแฟร์เพลย์เพื่อตรวจสอบแล้ว',
+  'fair_play.report_sent_short': 'ส่งรายงานแล้ว',
+  'fair_play.report_failed': 'ส่งรายงานแฟร์เพลย์ไม่สำเร็จ',
+  'fair_play.admin_title': 'เคสแฟร์เพลย์',
+  'fair_play.admin_desc': 'ตรวจสอบรายงานเกมจัดอันดับที่น่าสงสัย และจำกัดเฉพาะหลังแอดมินยืนยันแล้วเท่านั้น',
+  'fair_play.admin_count': '{count} เคส',
+  'fair_play.admin_empty': 'ไม่มีเคสแฟร์เพลย์ตามตัวกรองนี้',
+  'fair_play.admin_load_failed': 'โหลดเคสแฟร์เพลย์ไม่สำเร็จ',
+  'fair_play.admin_action_failed': 'อัปเดตเคสแฟร์เพลย์ไม่สำเร็จ',
+  'fair_play.admin_note_default': 'ตรวจสอบโดยแอดมิน',
+  'fair_play.filter_all': 'ทั้งหมด',
+  'fair_play.filter_open': 'เปิดอยู่',
+  'fair_play.filter_reviewed': 'ตรวจแล้ว',
+  'fair_play.filter_restricted': 'จำกัดแล้ว',
+  'fair_play.filter_dismissed': 'ยกเลิกแล้ว',
+  'fair_play.action_restrict': 'ปิดเกมจัดอันดับ',
+  'fair_play.action_clear': 'ล้างการจำกัด',
+  'fair_play.action_dismiss': 'ยกเลิกเคส',
+  'fair_play.event_count': '{count} สัญญาณ',
+  'fair_play.updated_at': 'อัปเดต',
+  'fair_play.case_id': 'เคส',
+  'fair_play.restricted_since': 'จำกัดตั้งแต่',
+  'fair_play.status.open': 'เปิดอยู่',
+  'fair_play.status.reviewed': 'ตรวจแล้ว',
+  'fair_play.status.restricted': 'จำกัดแล้ว',
+  'fair_play.status.dismissed': 'ยกเลิกแล้ว',
+  'fair_play.event.analysis_blocked': 'บล็อกการวิเคราะห์',
+  'fair_play.event.user_reported': 'ผู้ใช้รายงาน',
+  'fair_play.event.none': 'ไม่มีเหตุการณ์',
+
+  // Time ago
+  'time.just_now': 'เมื่อสักครู่',
+  'time.min_ago': '{n} นาทีที่แล้ว',
+  'time.hour_ago': '{n} ชั่วโมงที่แล้ว',
+  'time.day_ago': '{n} วันที่แล้ว',
+
+  // Analysis
+  'analysis.title': 'วิเคราะห์แบบเร็ว',
+  'analysis.loading': 'กำลังโหลดเกม...',
+  'analysis.parse_failed': 'อ่านข้อมูลเกมไม่สำเร็จ',
+  'analysis.game_not_found': 'ไม่พบเกม',
+  'analysis.no_moves': 'เกมนี้ไม่มีตาเดินให้วิเคราะห์',
+  'analysis.failed': 'วิเคราะห์ไม่สำเร็จ',
+  'analysis.analyzing': 'กำลังตรวจเกม...',
+  'analysis.analyzing_note': 'กำลังให้ Fairy-Stockfish ตรวจเกมนี้อยู่ เมื่อเอนจินประมวลผลเสร็จ รายงานจะขึ้นที่นี่',
+  'analysis.progress': 'ตาที่ {current} จาก {total}',
+  'analysis.elapsed': 'ใช้เวลาแล้ว {seconds} วินาที',
+  'analysis.slow_hint': 'เกมยาวอาจใช้เวลานานขึ้นบนเซิร์ฟเวอร์ปัจจุบัน รายงานจะอัปเดตต่อเมื่อวิเคราะห์ตำแหน่งถัดไปเสร็จ',
+  'analysis.engine_title': 'แหล่งที่มาของรีวิว',
+  'analysis.engine_note': 'รีวิวนี้ประมวลผลด้วย {engine}',
+  'analysis.provisional_title': 'รีวิวชั่วคราว',
+  'analysis.provisional_note': 'รีวิวนี้ต้องใช้การวิเคราะห์สำรองบางตำแหน่ง ({engine}) ทำให้ค่าความแม่นยำอาจยังสูงเกินจริง',
+  'analysis.accuracy': 'ความแม่นยำโดยประมาณ',
+  'analysis.eval_graph': 'ค่าประเมิน',
+  'analysis.show_best': 'แสดงตาแนะนำ',
+  'analysis.keyboard_hint': 'ใช้ลูกศรเพื่อดูตาเดิน',
+  'analysis.eval_before': 'ก่อน',
+  'analysis.eval_after': 'หลัง',
+  'analysis.best_was': 'ตาแนะนำคือ',
+  'analysis.best': 'ดีที่สุด',
+  'analysis.excellent': 'ยอดเยี่ยม',
+  'analysis.good': 'ดี',
+  'analysis.inaccuracy': 'ไม่แม่นยำ',
+  'analysis.mistake': 'ผิดพลาด',
+  'analysis.blunder': 'พลาดร้ายแรง',
+  'analysis.analyze': 'วิเคราะห์เกม',
+  'analysis.view': 'ดูการวิเคราะห์',
+  'analysis.next': 'ถัดไป',
+  'analysis.prev': 'ย้อนกลับ',
+  'analysis.note': 'รีวิวนี้ใช้เอนจินประเมินคุณภาพตาจากโอกาสชนะที่เปลี่ยนไป คล้ายแนวทางการคำนวณ accuracy ของ Lichess',
+  'analysis.move_accuracy': 'ความแม่นยำของตานี้',
+  'analysis.desc_best': 'ตานี้ตรงกับแนวเดินที่เอนจินมองว่าแข็งแรงที่สุด',
+  'analysis.desc_good': 'เป็นตาที่ดีและยังรักษาโอกาสชนะไว้ได้เกือบทั้งหมด',
+  'analysis.desc_inaccuracy': 'ตานี้ทำให้โอกาสชนะลดลงบ้าง แต่รูปยังเล่นต่อได้',
+  'analysis.desc_mistake': 'ตานี้ทำให้โอกาสชนะลดลงพอสมควร และเอนจินพบแนวที่ดีกว่า',
+  'analysis.desc_blunder': 'ตานี้ทำให้โอกาสชนะหายไปมาก และเอนจินพบแนวที่ดีกว่าชัดเจน',
+  'analysis.editor.error': 'วิเคราะห์ตำแหน่งไม่สำเร็จ',
+  'analysis.editor.reset_board': 'รีเซ็ตกระดาน',
+  'analysis.editor.clear_board': 'ล้างกระดาน',
+  'analysis.editor.analyzing_position': 'กำลังวิเคราะห์...',
+  'analysis.editor.analyze_position': 'วิเคราะห์ตำแหน่ง',
+  'analysis.editor.tools': 'เครื่องมือจัดตำแหน่ง',
+  'analysis.editor.move_pieces': 'ย้ายตัวหมาก',
+  'analysis.editor.erase_square': 'ลบช่อง',
+  'analysis.editor.position': 'ตำแหน่ง',
+  'analysis.editor.copy_position': 'คัดลอกตำแหน่ง',
+  'analysis.editor.copy_link': 'คัดลอกลิงก์',
+  'analysis.editor.engine': 'เอนจิน',
+  'analysis.editor.eval': 'การประเมิน',
+  'analysis.editor.best_move': 'ตาที่ดีที่สุด',
+  'analysis.editor.none': 'ไม่มี',
+  'analysis.editor.source': 'แหล่งที่มา',
+  'analysis.editor.depth': 'ความลึก',
+  'analysis.editor.pv': 'แนวหลัก',
+  'analysis.editor.label': 'ตัวแก้ตำแหน่ง',
+  'analysis.editor.turn_to_move': 'ถึงตาฝั่ง{color}',
+
+  // Auth / Account
+  'auth.sign_in': 'เข้าสู่ระบบ',
+  'auth.sign_in_desc': 'บัญชีผู้ใช้เป็นทางเลือก คุณยังเล่นแบบไม่ระบุตัวตนได้',
+  'auth.hero_title': 'เก็บเรตติ้งและโปรไฟล์ของคุณไว้กับบัญชีเดียว',
+  'auth.hero_desc': 'ใช้รหัสครั้งเดียวที่ส่งไปทางอีเมล คุณยังเล่นแบบสบาย ๆ โดยไม่ต้องมีบัญชีได้ แต่การเข้าสู่ระบบจะช่วยเก็บเรตติ้ง ชื่อผู้ใช้ และประวัติเกมไว้กับคุณ',
+  'auth.benefit_optional': 'ยังเล่นแบบไม่ระบุตัวตนได้สำหรับเกมสบาย ๆ',
+  'auth.benefit_rated': 'เกมจัดอันดับและการตั้งค่าบัญชีจะผูกกับอีเมลของคุณ',
+  'auth.benefit_fast': 'ไม่ต้องจำรหัสผ่าน แค่อีเมล รหัส แล้วเข้าเล่นได้เลย',
+  'auth.back_to_play': 'กลับไปเล่น',
+  'auth.step_email': 'อีเมล',
+  'auth.step_code': 'รหัส',
+  'auth.email': 'อีเมล',
+  'auth.email_placeholder': 'you@example.com',
+  'auth.email_hint': 'เราจะส่งรหัส 6 หลักแบบใช้ครั้งเดียวไปที่อีเมลนี้',
+  'auth.code': 'รหัส 6 หลัก',
+  'auth.code_placeholder': '123456',
+  'auth.code_hint': 'โดยปกติรหัสจะมาถึงภายใน 1 นาที ถ้ายังไม่เห็นให้ตรวจสอบโฟลเดอร์สแปม',
+  'auth.send_code': 'ส่งรหัสเข้าสู่ระบบ',
+  'auth.sending_code': 'กำลังส่งรหัส...',
+  'auth.code_sent_notice': 'ตรวจสอบอีเมลของคุณเพื่อรับรหัสเข้าสู่ระบบ 6 หลัก',
+  'auth.code_sent_to': 'ส่งรหัสไปที่ {email}',
+  'auth.verify_code': 'ยืนยันรหัส',
+  'auth.signing_in': 'กำลังเข้าสู่ระบบ...',
+  'auth.use_another_email': 'ใช้อีเมลอื่น',
+  'auth.send_code_failed': 'ส่งรหัสไม่สำเร็จ',
+  'auth.sign_in_failed': 'เข้าสู่ระบบไม่สำเร็จ',
+  'account.title': 'บัญชี',
+  'account.username': 'ชื่อผู้ใช้',
+  'account.username_placeholder': '3-20 ตัวอักษร ตัวเลข หรือขีดล่าง',
+  'account.rating': 'เรตติ้ง',
+  'account.rated_games': 'เกมจัดอันดับ',
+  'account.wins': 'ชนะ',
+  'account.losses': 'แพ้',
+  'account.draws': 'เสมอ',
+  'account.saving': 'กำลังบันทึก...',
+  'account.save_profile': 'บันทึกโปรไฟล์',
+  'account.profile_updated': 'อัปเดตโปรไฟล์แล้ว',
+  'account.update_failed': 'อัปเดตโปรไฟล์ไม่สำเร็จ',
+  'account.open_feedback': 'เปิดหน้าจัดการความคิดเห็น',
+  'account.open_fair_play': 'เปิดหน้าจัดการแฟร์เพลย์',
+  'account.rated_restricted_title': 'จำกัดเกมจัดอันดับ',
+  'account.rated_restricted_desc': 'บัญชีนี้ยังเข้าสู่ระบบและเล่นเกมทั่วไปได้ แต่ปิดการจับคู่แบบจัดอันดับไว้ระหว่างรอตรวจสอบแฟร์เพลย์',
+  'account.sign_out': 'ออกจากระบบ',
+  'account.puzzle_eyebrow': 'ความคืบหน้าปริศนา',
+  'account.puzzle_title': 'ภาพรวมการฝึก',
+  'account.puzzle_desc': 'ประวัติปริศนาของคุณจะติดไปกับบัญชีและกลับมาได้ทุกอุปกรณ์',
+  'account.puzzle_completed_label': 'แก้แล้ว',
+  'account.puzzle_remaining_label': 'เหลือ',
+  'account.puzzle_focus_label': 'ธีมเด่น',
+  'account.puzzle_focus_empty': 'ยังไม่มีธีม',
+  'account.puzzle_next_label': 'เล่นต่อที่',
+  'account.puzzle_continue': 'ฝึกต่อ',
+  'account.puzzle_all_done': 'คุณเคลียร์ชุดปัจจุบันแล้ว',
+  'account.puzzle_all_done_desc': 'เปิดดูรายการทั้งหมดอีกครั้งและทบทวนโจทย์ที่ยากที่สุด',
+  'account.puzzle_last_played_label': 'เล่นล่าสุด',
+  'account.puzzle_last_played_empty': 'เริ่มเล่นปริศนาแล้วมันจะมาแสดงตรงนี้',
+  'account.puzzle_last_played_meta': '{status} · {date}',
+  'account.puzzle_recent_label': 'ปริศนาที่เพิ่งแก้',
+  'account.puzzle_recent_empty': 'ยังไม่มีปริศนาที่แก้สำเร็จ',
+  'account.puzzle_recent_meta': 'แก้เมื่อ {date}',
+  'account.puzzle_status_in_progress': 'กำลังเล่น',
+  'account.puzzle_status_solved': 'แก้แล้ว',
+  'account.puzzle_percent': 'สำเร็จแล้ว {percent}% ของชุดปัจจุบัน {total} ปริศนา',
+
+  // Leaderboard
+  'leaderboard.eyebrow': 'อันดับเกมจัดอันดับ',
+  'leaderboard.title': 'ตารางคะแนน',
+  'leaderboard.desc': 'ผู้เล่นเรตติ้งสูงสุดจากเกมควิกเพลย์แบบจัดอันดับที่จบแล้ว',
+  'leaderboard.count': 'ผู้เล่นติดอันดับ {count} คน',
+  'leaderboard.empty': 'ยังไม่มีผู้เล่นจัดอันดับ',
+  'leaderboard.empty_desc': 'เล่นควิกเพลย์แบบจัดอันดับให้จบสักเกม แล้วชื่อจะขึ้นที่นี่',
+  'leaderboard.col_rank': 'อันดับ',
+  'leaderboard.col_player': 'ผู้เล่น',
+  'leaderboard.col_rating': 'เรตติ้ง',
+  'leaderboard.col_games': 'เกม',
+  'leaderboard.col_record': 'สถิติ',
+  'leaderboard.col_winrate': 'อัตราชนะ',
+  'leaderboard.you': 'คุณ',
+  'leaderboard.view_recent': 'เกมล่าสุด',
+  'leaderboard.mobile_stats': '{games} เกม • {record} • {winrate}',
+  'leaderboard.load_failed': 'โหลดตารางคะแนนไม่สำเร็จ ({status})',
+
+  // Error UI
+  'error.board_display': 'เกิดข้อผิดพลาดในการแสดงกระดาน',
+  'error.connection_title': 'ข้อผิดพลาดการเชื่อมต่อ',
+  'error.connection_body': 'เกิดข้อผิดพลาดขณะโหลดข้อมูล',
+  'error.try_again': 'ลองอีกครั้ง',
+  'error.something_wrong': 'เกิดข้อผิดพลาดบางอย่าง',
+  'error.unexpected': 'แอปพบข้อผิดพลาดที่ไม่คาดคิด และได้บันทึกไว้แล้ว',
+  'error.unknown': 'ข้อผิดพลาดที่ไม่ทราบสาเหตุ',
+  'error.reload_page': 'โหลดหน้าใหม่',
+  'error.report_bug': 'รายงานบั๊กนี้บน GitHub',
+
+  // Header
+  'header.admin': 'แอดมิน',
+  'header.sign_in': 'เข้าสู่ระบบ',
+  'header.menu': 'เมนู',
+  'header.close_menu': 'ปิด',
+  'header.switch_to_th': 'เปลี่ยนเป็นภาษาไทย',
+  'header.switch_to_en': 'เปลี่ยนเป็นภาษาอังกฤษ',
+
+  // Guide Page
+  'guide.cta_title': 'พร้อมลองเล่นหรือยัง',
+  'guide.cta_desc': 'ถ้าอยากลองทันที ให้เริ่มจากบอทหรือโจทย์ก่อน แล้วค่อยไปเล่นกับคนจริงเมื่อเริ่มคุ้นกับเกม',
+  'guide.cta_next': 'ไปหน้าถัดไป',
+
+  // Footer
+  'footer.links_label': 'ลิงก์ส่วนท้ายเว็บไซต์',
+  'footer.what_is_makruk': 'หมากรุกไทยคืออะไร',
+  'footer.how_to_play_makruk': 'วิธีเล่นหมากรุกไทย',
+
+  // Language
+  'lang.switch': 'EN',
+};
