+++ client/src/test/i18nCatalog.test.ts
@@ -1,5 +1,6 @@
 import { describe, expect, it } from 'vitest';
-import { EN_TRANSLATIONS, TH_TRANSLATIONS } from '../lib/i18n';
+import { EN_TRANSLATIONS } from '../lib/i18n';
+import { TH_TRANSLATIONS } from '../lib/i18n.th';
 
 function findMissingKeys(
   source: Record<string, string>,
@@ -26,7 +27,7 @@ describe('i18n catalogs', () => {
       { missingInThai, missingInEnglish },
       [
         'Translation catalogs drifted.',
-        'When you add a new UI string, add both English and Thai entries in client/src/lib/i18n.tsx.',
+        'When you add a new UI string, add both English and Thai entries in client/src/lib/i18n.tsx and client/src/lib/i18n.th.ts.',
         `Missing in Thai: ${missingInThai.join(', ') || 'none'}`,
         `Missing in English: ${missingInEnglish.join(', ') || 'none'}`,
       ].join('\n'),
